// generated/src/app/api-models/pmb-sub-det-memb-vol-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbSubDetMembVolWork {
  seqGpbilId: number;
  seqSubDtlId: number;
  salary: number;
  memberVolume: number;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}